move unit test functionality
